parallel case

+++ tests/milvus_python_test/conftest.py
@@ -133,8 +133,7 @@ def collection(request, connect):
     except Exception as e:
         pytest.exit(str(e))
     def teardown():
-        collection_names = connect.list_collections()
-        for collection_name in collection_names:
+        if connect.has_collection(collection_name):
             connect.drop_collection(collection_name, timeout=delete_timeout)
     request.addfinalizer(teardown)
     assert connect.has_collection(collection_name)
@@ -152,8 +151,7 @@ def id_collection(request, connect):
     except Exception as e:
         pytest.exit(str(e))
     def teardown():
-        collection_names = connect.list_collections()
-        for collection_name in collection_names:
+        if connect.has_collection(collection_name):
             connect.drop_collection(collection_name, timeout=delete_timeout)
     request.addfinalizer(teardown)
     assert connect.has_collection(collection_name)
@@ -170,8 +168,7 @@ def binary_collection(request, connect):
     except Exception as e:
         pytest.exit(str(e))
     def teardown():
-        collection_names = connect.list_collections()
-        for collection_name in collection_names:
+        if connect.has_collection(collection_name):
             connect.drop_collection(collection_name, timeout=delete_timeout)
     request.addfinalizer(teardown)
     assert connect.has_collection(collection_name)
@@ -189,8 +186,7 @@ def binary_id_collection(request, connect):
     except Exception as e:
         pytest.exit(str(e))
     def teardown():
-        collection_names = connect.list_collections()
-        for collection_name in collection_names:
+        if connect.has_collection(collection_name):
             connect.drop_collection(collection_name, timeout=delete_timeout)
     request.addfinalizer(teardown)
     assert connect.has_collection(collection_name)
